+++ src/C++/Dictionary.h
@@ -53,7 +53,7 @@ public:
   std::string getString( const std::string&, bool capitalize = false ) const
   throw( ConfigError, FieldConvertError );
   /// Get a value as a int.
-  int getLong( const std::string& ) const
+  int getInt( const std::string& ) const
   throw( ConfigError, FieldConvertError );
   /// Get a value as a double.
   double getDouble( const std::string& ) const
@@ -68,7 +68,7 @@ public:
   /// Set a value from a string.
   void setString( const std::string&, const std::string& );
   /// Set a value from a int.
-  void setLong( const std::string&, int );
+  void setInt( const std::string&, int );
   /// Set a value from a double.
   void setDouble( const std::string&, double );
   /// Set a value from a bool
