+++ src/C++/HttpServer.cpp
@@ -69,7 +69,7 @@ HttpServer::HttpServer( const SessionSettings& settings ) throw( ConfigError )
 void HttpServer::onConfigure( const SessionSettings& s )
 throw ( ConfigError )
 {  
-  m_port = s.get().getLong( HTTP_ACCEPT_PORT );
+  m_port = s.get().getInt( HTTP_ACCEPT_PORT );
 }
 
 void HttpServer::onInitialize( const SessionSettings& s )
